models/tests/test_robustness.py
"""Tests for robustness infrastructure.
Covers artifact I/O, noise injection correctness, comparison table schema,
tarpit handler format, and concurrent session ID uniqueness.
"""
import json
from pathlib import Path
import numpy as np
import pandas as pd
import pytest
from api.app.config import PROJECT_ROOT, RANDOM_SEED
from models.robustness import (
    ARTIFACT_PATHS,
    SIGMA_LEVELS,
    TIMING_NOISE_COLS,
    inject_timing_noise,
    load_artifacts,
)
from models.classifiers import (
    ALL_FEATURES,
    ALL_FEATURES_LR,
    load_features,
    split_data,
    evaluate_model,
)

# Paths
FEATURES_CSV   = PROJECT_ROOT / "data" / "processed" / "features.csv"
ARTIFACTS_DIR  = PROJECT_ROOT / "models" / "artifacts"
METRICS_DIR    = PROJECT_ROOT / "report" / "metrics"

# Required keys in the final robustness comparison JSON
REQUIRED_EXPERIMENT_KEYS = {"baseline", "tarpit", "concurrent", "ood_model", "ood_prompt"}


#  Fixtures 

@pytest.fixture(scope="module")
def features_df():
    """Load the features CSV once for the whole module."""
    if not FEATURES_CSV.exists():
        pytest.skip("features.csv not found — run the Phase 3 pipeline first")
    return load_features(FEATURES_CSV)


@pytest.fixture(scope="module")
def artifacts():
    """Load saved model artifacts once for the whole module."""
    missing = [k for k, p in ARTIFACT_PATHS.items() if not p.exists()]
    if missing:
        pytest.skip(f"Artifacts not found: {missing}. Run --save-artifacts first.")
    return load_artifacts()


# Test 1: Artifact save and load produces identical predictions

def test_artifact_save_load(features_df, artifacts):
    """RF loaded from disk produces same predictions as a freshly trained RF."""
    from models.robustness import _apply_preprocessors
    df = features_df
    le  = artifacts["label_enc"]
    rf  = artifacts["rf"]
    y   = le.transform(df["actor_type"])

    X_rf = _apply_preprocessors(df, ALL_FEATURES, artifacts["imputer_rf"], artifacts["scaler_rf"])

    # Predictions from disk-loaded model
    preds_loaded = rf.predict(X_rf)

    # Retrain fresh model with same seed and same split, compare predictions
    train_df, test_df, y_train, y_test, le2 = split_data(df, seed=RANDOM_SEED)
    from models.robustness import _apply_preprocessors as ap
    from sklearn.impute import SimpleImputer
    from sklearn.preprocessing import StandardScaler
    from models.classifiers import make_rf

    X_tr = train_df[ALL_FEATURES].copy()
    nan_cols = [c for c in ALL_FEATURES if X_tr[c].isna().any()]
    for col in nan_cols:
        X_tr[f"{col}_missing"] = X_tr[col].isna().astype(int)
    imp = SimpleImputer(strategy="median").fit(X_tr)
    scl = StandardScaler().fit(imp.transform(X_tr))
    rf2 = make_rf(seed=RANDOM_SEED)
    rf2.fit(scl.transform(imp.transform(X_tr)), y_train)

    X_all = ap(df, ALL_FEATURES, imp, scl)
    preds_fresh = rf2.predict(X_all)

    assert np.array_equal(preds_loaded, preds_fresh), (
        "Disk-loaded RF predictions differ from freshly trained RF — "
        "artifact may be stale or seed mismatch"
    )


# Test 2: Preprocessors transform without refitting

def test_preprocessor_no_refit(features_df, artifacts):
    """Imputer and scaler loaded from disk are used with transform(), not fit_transform().
    Verifies that _apply_preprocessors never calls fit on the loaded objects."""
    from models.robustness import _apply_preprocessors
    from unittest.mock import patch
    imp = artifacts["imputer_rf"]
    scl = artifacts["scaler_rf"]

    # Patch fit_transform to raise. if it's called, the test fails
    with patch.object(type(imp), "fit_transform", side_effect=AssertionError("fit_transform called on imputer")):
        with patch.object(type(scl), "fit_transform", side_effect=AssertionError("fit_transform called on scaler")):
            X = _apply_preprocessors(features_df, ALL_FEATURES, imp, scl)

    assert X.shape[0] == len(features_df)
    assert X.shape[1] > 0


# Test 3: Noise injection never produces negative values 

def test_noise_injection_nonnegative(features_df):
    """After noise injection at any sigma level, all timing features must be >= 0."""
    for sigma in SIGMA_LEVELS:
        noisy = inject_timing_noise(features_df, sigma_ms=sigma, seed=RANDOM_SEED)
        for col in TIMING_NOISE_COLS:
            if col not in noisy.columns:
                continue
            values = noisy[col].dropna()
            assert (values >= 0).all(), (
                f"Negative values found in '{col}' at sigma={sigma}ms"
            )


# Test 4: Zero noise reproduces Phase 5 F1 on full dataset 

def test_noise_zero_reproduces_baseline(features_df, artifacts):
    """inject_timing_noise with sigma=0 is a no-op — F1 must match sigma=0 baseline."""
    from models.robustness import _apply_preprocessors
    le = artifacts["label_enc"]
    rf = artifacts["rf"]
    y  = le.transform(features_df["actor_type"])

    # Clean data
    X_clean = _apply_preprocessors(features_df, ALL_FEATURES, artifacts["imputer_rf"], artifacts["scaler_rf"])
    _, report_clean, _ = evaluate_model(rf, X_clean, y, list(le.classes_))
    f1_clean = report_clean["macro avg"]["f1-score"]

    # Zero-noise data 
    noisy_df = inject_timing_noise(features_df, sigma_ms=0, seed=RANDOM_SEED)
    X_noisy  = _apply_preprocessors(noisy_df, ALL_FEATURES, artifacts["imputer_rf"], artifacts["scaler_rf"])
    _, report_noisy, _ = evaluate_model(rf, X_noisy, y, list(le.classes_))
    f1_noisy = report_noisy["macro avg"]["f1-score"]

    assert abs(f1_clean - f1_noisy) < 1e-6, (
        f"sigma=0 changed F1: clean={f1_clean:.6f}, noisy={f1_noisy:.6f}"
    )


# Test 5: Comparison table has required experiment keys 

def test_comparison_table_keys():
    """robustness_results.json must contain all expected experiment keys when it exists.
    Skips gracefully if the file hasn't been written yet (pre-experiment runs)."""
    results_path = METRICS_DIR / "robustness_results.json"
    if not results_path.exists():
        pytest.skip("robustness_results.json not yet written — run all experiments first")

    with open(results_path, "r", encoding="utf-8") as f:
        data = json.load(f)

    missing_keys = REQUIRED_EXPERIMENT_KEYS - set(data.keys())
    assert not missing_keys, (
        f"robustness_results.json missing experiment keys: {missing_keys}"
    )


#  Test 6: Tarpit handler returns 200 

def test_tarpit_handler_returns_200():
    """The tarpit handler function returns status 200 with a success body.
       Tests the handler logic directly without needing a live server."""
    import asyncio
    from unittest.mock import MagicMock
    from slowapi.errors import RateLimitExceeded

    # Import the tarpit handler. it lives in main.py as _rate_limit_handler.
    # its *design contract* is verified by calling it directly.
    # Note: when tarpitting is reverted to honest 429, this test documents
    # what the tarpit version returned during the previous experiment.
    # The assertion checks the metrics file produced by that experiment instead.
    results_path = PROJECT_ROOT / "report" / "metrics" / "robustness_tarpit.json"
    if not results_path.exists():
        pytest.skip("robustness_tarpit.json not found — run Step 6.2 first")

    with open(results_path, "r", encoding="utf-8") as f:
        data = json.load(f)

    # Tarpit experiment must have run and produced a valid F1 score
    assert "rf_f1" in data, "robustness_tarpit.json missing rf_f1 key"
    assert isinstance(data["rf_f1"], float), "rf_f1 must be a float"
    assert data["rf_f1"] > 0.90, (
        f"Tarpit RF F1 {data['rf_f1']:.4f} fell below 0.90 robustness threshold"
    )


#  Test 7: Concurrent simulation produces unique session IDs 

def test_concurrent_session_ids_unique():
    """Concurrent telemetry JSONL must have no duplicate session_id values.
    Skips gracefully if concurrent output doesn't exist yet."""
    concurrent_jsonl = PROJECT_ROOT / "data" / "raw" / "telemetry_concurrent.jsonl"
    if not concurrent_jsonl.exists():
        pytest.skip("telemetry_concurrent.jsonl not yet generated")

    import json as _json
    session_ids = []
    with open(concurrent_jsonl, "r", encoding="utf-8") as f:
        for line in f:
            line = line.strip()
            if line:
                record = _json.loads(line)
                session_ids.append(record["session_id"])

    # session_id appears once per request, duplicates are expected within a session
    # What we're checking: no two *different* sessions share the same ID
    # A session ID appears N times where N = request count for that session
    # So uniqueness at the session level is checked by grouping
    from collections import Counter
    # Load features CSV to get session-level actor_type mapping
    concurrent_features = PROJECT_ROOT / "data" / "processed" / "features_concurrent.csv"
    if not concurrent_features.exists():
        pytest.skip("features_concurrent.csv not yet generated — run pipeline after the step is completed")

    df = pd.read_csv(concurrent_features)
    # Each session_id must map to exactly one actor_type
    actor_per_session = df.groupby("session_id")["actor_type"].nunique()
    mixed_sessions = actor_per_session[actor_per_session > 1]
    assert len(mixed_sessions) == 0, (
        f"{len(mixed_sessions)} session(s) have mixed actor_types — "
        "session IDs may have collided across concurrent simulators"
    )
